requirements.txt
python-telegram-bot==21.9
pytest==8.3.4
pydantic==2.10.4